Splitting files up within script-viewer.
